+++ src/geo/geometry-models/polygon.js
@@ -1,0 +1,10 @@
+var PathBase = require('./path-base');
+
+var Polygon = PathBase.extend({
+  defaults: {
+    type: 'polygon',
+    latlngs: []
+  }
+});
+
+module.exports = Polygon;
