+++ src/geo/geometry-models/polyline.js
@@ -1,0 +1,10 @@
+var PathBase = require('./path-base');
+
+var Polyline = PathBase.extend({
+  defaults: {
+    type: 'polyline',
+    latlngs: []
+  }
+});
+
+module.exports = Polyline;
