+++ src/geo/leaflet/geometries/polygon-view.js
@@ -1,0 +1,10 @@
+var L = require('leaflet');
+var PathViewBase = require('./path-view-base');
+
+var PolygonView = PathViewBase.extend({
+  _createGeometry: function () {
+    return L.polygon([], { color: 'red' });
+  }
+});
+
+module.exports = PolygonView;
