+++ src/geo/leaflet/geometries/polyline-view.js
@@ -1,0 +1,10 @@
+var L = require('leaflet');
+var PathViewBase = require('./path-view-base');
+
+var PolylineView = PathViewBase.extend({
+  _createGeometry: function () {
+    return L.polyline([], { color: 'red' });
+  }
+});
+
+module.exports = PolylineView;
